requirements.txt
numpy
pytorch
cudatoolkit
torchvision
torchaudio
matplotlib
